Change heartbeat from u64 to u8 since it is within that range only

+++ shared/src/config.rs
@@ -15,7 +15,7 @@ pub struct Server {
 
 #[derive(Debug, Deserialize)]
 pub struct Worker {
-    pub heartbeat: u64,
+    pub heartbeat: u8,
 }
 
 #[derive(Debug, Deserialize)]
